Kinesis/DynamoDB: Refactor KCLv2 implementation to `lorrystream.spike`
It needs further curation. The next iteration should aim towards
slotting it in as a native streamz Source element.
