+++ lorrystream/spike/kcl_dynamodb/dynamodb_cdc_processor.py
@@ -5,32 +5,41 @@
 
 from __future__ import print_function
 
+import json
 import logging
 import logging.handlers as handlers
+import os
 import time
 import typing as t
 
 from amazon_kclpy import kcl
 from amazon_kclpy.v3 import processor
+from commons_codec.transform.dynamodb import DynamoCDCTranslatorCrateDB
+from cratedb_toolkit.util import DatabaseAdapter
 
-# Logger writes to file because stdout is used by MultiLangDaemon
 logger = logging.getLogger(__name__)
-logger.setLevel(logging.INFO)
-formatter = logging.Formatter(
-    "%(asctime)s.%(msecs)03d [%(module)s] %(levelname)s  %(funcName)s - %(message)s", "%H:%M:%S"
-)
-handler = handlers.RotatingFileHandler("./record_processor.log", maxBytes=10**6, backupCount=5)
-handler.setLevel(logging.INFO)
-handler.setFormatter(formatter)
-logger.addHandler(handler)
-
 
 IntOrNone = t.Union[int, None]
+FloatOrNone = t.Union[float, None]
+
+
+def setup_logging(logfile: str):
+    """
+    Configure Python logger to write to file, because stdout is used by MultiLangDaemon.
+    """
+    logger.setLevel(logging.INFO)
+    formatter = logging.Formatter(
+        "%(asctime)s.%(msecs)03d [%(module)s] %(levelname)s  %(funcName)s - %(message)s", "%H:%M:%S"
+    )
+    handler = handlers.RotatingFileHandler(logfile, maxBytes=10**6, backupCount=5)
+    handler.setLevel(logging.INFO)
+    handler.setFormatter(formatter)
+    logger.addHandler(handler)
 
 
 class RecordProcessor(processor.RecordProcessorBase):
     """
-    A RecordProcessor processes data from a shard in a stream. Its methods will be called with this pattern:
+    Process data from a shard in a stream. Its methods will be called with this pattern:
 
     * initialize will be called once
     * process_records will be called zero or more times
@@ -38,13 +47,26 @@ class RecordProcessor(processor.RecordProcessorBase):
         a scaling change.
     """
 
-    def __init__(self):
+    def __init__(self, sqlalchemy_url: t.Optional[str], table_name: t.Optional[str]):
         self._SLEEP_SECONDS = 5
         self._CHECKPOINT_RETRIES = 5
         self._CHECKPOINT_FREQ_SECONDS = 60
         self._largest_seq: t.Tuple[IntOrNone, IntOrNone] = (None, None)
         self._largest_sub_seq = None
-        self._last_checkpoint_time = None
+        self._last_checkpoint_time: FloatOrNone = None
+
+        self.sqlalchemy_url = sqlalchemy_url
+        self.table_name = table_name
+
+        # Sanity checks.
+        if self.sqlalchemy_url is None:
+            raise ValueError("SQLAlchemy URL must not be empty")
+        if self.table_name is None:
+            raise ValueError("Target CDC table name must not be empty")
+
+        self.cratedb = DatabaseAdapter(dburi=self.sqlalchemy_url)
+        self.table_name = self.table_name
+        self.cdc = DynamoCDCTranslatorCrateDB(table_name=self.table_name)
 
     def initialize(self, initialize_input):
         """
@@ -99,18 +121,32 @@ class RecordProcessor(processor.RecordProcessorBase):
 
     def process_record(self, data, partition_key, sequence_number, sub_sequence_number):
         """
-        Called for each record that is passed to process_records.
+        Convert record, which is a DynamoDB CDC event item, into an SQL statement,
+        and submit to downstream database.
 
         :param str data: The blob of data that was contained in the record.
         :param str partition_key: The key associated with this recod.
         :param int sequence_number: The sequence number associated with this record.
         :param int sub_sequence_number: the sub sequence number associated with this record.
         """
-        ####################################
-        # Insert your processing logic here
-        ####################################
 
-        logger.info(data.decode("UTF-8"))
+        sql = None
+        try:
+            cdc_event = json.loads(data)
+            logger.info("CDC event: %s", cdc_event)
+
+            sql = self.cdc.to_sql(cdc_event)
+            logger.info("SQL: %s", sql)
+        except Exception:
+            logger.exception("Decoding CDC event failed")
+
+        if not sql:
+            return
+
+        try:
+            self.cratedb.run_sql(sql)
+        except Exception:
+            logger.exception("Writing CDC event to sink database failed")
 
     def should_update_sequence(self, sequence_number, sub_sequence_number):
         """
@@ -166,6 +202,20 @@ class RecordProcessor(processor.RecordProcessorBase):
         shutdown_requested_input.checkpointer.checkpoint()
 
 
-if __name__ == "__main__":
-    kcl_process = kcl.KCLProcess(RecordProcessor())
+def main():
+    # Set up logging.
+    logfile = os.environ.get("CDC_LOGFILE", "cdc.log")
+    setup_logging(logfile)
+
+    # Setup processor.
+    sqlalchemy_url = os.environ.get("CDC_SQLALCHEMY_URL")
+    table_name = os.environ.get("CDC_TABLE_NAME")
+    kcl_processor = RecordProcessor(sqlalchemy_url=sqlalchemy_url, table_name=table_name)
+
+    # Invoke machinery.
+    kcl_process = kcl.KCLProcess(kcl_processor)
     kcl_process.run()
+
+
+if __name__ == "__main__":
+    main()
