+++ lorrystream/spike/kcl_dynamodb/launch.sh
