+++ lorrystream/spike/kcl_dynamodb/requirements.txt
@@ -1,4 +1,3 @@
 amazon-kclpy==2.1.5
 awscli==1.33.*
 boto3<1.35
-simplejson<4
