+++ lorrystream/spike/kcl_kinesis/.gitignore
