+++ lorrystream/spike/kcl_kinesis/launch.sh
