+++ lorrystream/spike/kcl_kinesis/record_processor.properties
@@ -29,7 +29,7 @@ initialPositionInStream = TRIM_HORIZON
 # by the MultiLangDaemon.
 
 # The KCL defaults to us-east-1, this value is overridden by the set_region.py scripts
-regionName = us-east-1
+regionName = eu-central-1
 
 # Fail over time in milliseconds. A worker which does not renew it's lease within this time interval
 # will be regarded as having problems and it's shards will be assigned to other workers.
