+++ lorrystream/spike/kcl_kinesis/requirements.txt
